Consolidate final platforms
