+++ homeassistant/components/demo/calendar.py
@@ -9,7 +9,7 @@ import copy
 from homeassistant.components.google import CONF_DEVICE_ID, CONF_NAME
 import homeassistant.util.dt as dt_util
 
-from . import CalendarEventDevice, get_date
+from homeassistant.components.calendar import CalendarEventDevice, get_date
 
 
 def setup_platform(hass, config, add_entities, discovery_info=None):
