+++ homeassistant/components/demo/water_heater.py
@@ -1,7 +1,7 @@
 """Demo platform that offers a fake water heater device."""
 from homeassistant.const import ATTR_TEMPERATURE, TEMP_CELSIUS, TEMP_FAHRENHEIT
 
-from . import (
+from homeassistant.components.water_heater import (
     SUPPORT_AWAY_MODE, SUPPORT_OPERATION_MODE, SUPPORT_TARGET_TEMPERATURE,
     WaterHeaterDevice)
