+++ homeassistant/components/generic_thermostat/__init__.py
@@ -1,0 +1,1 @@
+"""The generic_thermostat component."""
