+++ homeassistant/components/manual/alarm_control_panel.py
