+++ homeassistant/components/manual_mqtt/alarm_control_panel.py
