+++ tests/components/demo/test_calendar.py
