+++ tests/components/demo/test_camera.py
