+++ tests/components/demo/test_climate.py
