+++ tests/components/demo/test_cover.py
