+++ tests/components/demo/test_fan.py
