+++ tests/components/demo/test_light.py
