+++ tests/components/demo/test_lock.py
