+++ tests/components/demo/test_media_player.py
