+++ tests/components/demo/test_water_heater.py
