+++ tests/components/generic_thermostat/test_climate.py
