+++ tests/components/google_assistant/test_smart_home.py
@@ -11,7 +11,7 @@ from homeassistant.components.climate.const import (
 from homeassistant.components.google_assistant import (
     const, trait, helpers, smart_home as sh,
     EVENT_COMMAND_RECEIVED, EVENT_QUERY_RECEIVED, EVENT_SYNC_RECEIVED)
-from homeassistant.components.light.demo import DemoLight
+from homeassistant.components.demo.light import DemoLight
 
 from homeassistant.helpers import device_registry
 from tests.common import (mock_device_registry, mock_registry,
