+++ tests/components/manual/test_alarm_control_panel.py
