+++ tests/components/manual_mqtt/test_alarm_control_panel.py
