+++ tests/components/switch/test_light.py
